- Fixed warnings

+++ Pod/Classes/Protocols/NSDictionary+DVATableViewModelDatasource.m
@@ -16,7 +16,7 @@
 -(NSArray*)viewModelsForSection:(NSInteger)section{
     NSArray*viewModels=[self objectForKey:@(section)];
     if (!viewModels) {
-        viewModels=[self objectForKey:[NSString stringWithFormat:@"%i",section]];
+        viewModels=[self objectForKey:[NSString stringWithFormat:@"%li",(long)section]];
     }
     return viewModels;
 }
